+++ src/requisition/requisitions.service.js
@@ -469,30 +469,35 @@
         }
 
         function extendLineItemsWithOrderablesAndFtaps(requisition, statusMessages) {
-            return $q.all([getByVersionIdentities(requisition.availableProducts, new OrderableResource()),
+            var identities = getResourcesFromLineItems(requisition, true);
+            return $q.all([getByVersionIdentities(identities, new OrderableResource()),
                 periodService.get(requisition.processingPeriod.id)])
                 .then(function(result) {
-                    requisition.availableFullSupplyProducts =
-                        filterOrderables(true, result[0], requisition.program.id);
-                    requisition.availableNonFullSupplyProducts =
-                        filterOrderables(false, result[0], requisition.program.id);
+                    requisition.requisitionLineItems.forEach(function(lineItem) {
+                        result[0].forEach(function(orderable) {
+                            if (lineItem.orderable.id === orderable.id) {
+                                lineItem.orderable = orderable;
+                            }
+                        });
+                    });
 
                     requisition.processingPeriod = result[1];
                     return requisition;
                 })
                 .then(function(requisition) {
-                    var indentities = getResourcesFromLineItems(requisition.requisitionLineItems);
-                    return $q.all([
-                        getByVersionIdentities(indentities.orderables, new OrderableResource()),
-                        getByVersionIdentities(indentities.ftaps, new FacilityTypeApprovedProductResource())
-                    ])
+                    var identities = getResourcesFromLineItems(requisition, false),
+                        fetchAvailableProducts = requisition.emergency ?
+                            getByVersionIdentities(requisition.availableProducts, new OrderableResource()) :
+                            $q.resolve([]);
+                    return $q.all([fetchAvailableProducts,
+                        getByVersionIdentities(identities, new FacilityTypeApprovedProductResource())])
                         .then(function(result) {
+                            requisition.availableFullSupplyProducts =
+                                filterOrderables(true, result[0], requisition.program.id);
+                            requisition.availableNonFullSupplyProducts =
+                                filterOrderables(false, result[0], requisition.program.id);
+
                             requisition.requisitionLineItems.forEach(function(lineItem) {
-                                result[0].forEach(function(orderable) {
-                                    if (lineItem.orderable.id === orderable.id) {
-                                        lineItem.orderable = orderable;
-                                    }
-                                });
                                 result[1].forEach(function(ftap) {
                                     if (lineItem.approvedProduct && lineItem.approvedProduct.id === ftap.id) {
                                         lineItem.approvedProduct = ftap;
@@ -504,20 +509,19 @@
                 });
         }
 
-        function getResourcesFromLineItems(lineItems) {
-            var orderableIdentities = [],
-                ftapIdentities = [];
-            lineItems.forEach(function(item) {
-                orderableIdentities.push(item.orderable);
-                if (item.approvedProduct) {
-                    ftapIdentities.push(item.approvedProduct);
+        function getResourcesFromLineItems(requisition, isOrderable) {
+            var identities = [];
+            requisition.requisitionLineItems.forEach(function(item) {
+                if (isOrderable) {
+                    identities.push(item.orderable);
+                } else {
+                    var program = getProgramById(item.orderable.programs, requisition.program.id);
+                    if (item.approvedProduct && program.fullSupply) {
+                        identities.push(item.approvedProduct);
+                    }
                 }
             });
-
-            return {
-                ftaps: ftapIdentities,
-                orderables: orderableIdentities
-            };
+            return identities;
         }
 
         function getByVersionIdentities(identities, resource) {